fix: move ffetch to lib-ffetch
